requirements.txt
transformers
torch
torchaudio
pydub
pyannote.audio
speechbrain